requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=dev.txt dev.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
bcrypt==5.0.0
    # via passlib
black==26.3.1
    # via -r dev.in
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via
    #   black
    #   uvicorn
coverage[toml]==7.13.4
    # via pytest-cov
cryptography==46.0.5
    # via python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.135.1
    # via -r base.in
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r dev.in
idna==3.11
    # via
    #   anyio
    #   httpx
iniconfig==2.3.0
    # via pytest
librt==0.8.1
    # via mypy
mypy==1.19.1
    # via -r dev.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==26.0
    # via
    #   black
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r base.in
pathspec==1.0.4
    # via
    #   black
    #   mypy
platformdirs==4.9.4
    # via black
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
psycopg2-binary==2.9.11
    # via -r base.in
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r base.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r base.in
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r dev.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==1.3.0
    # via -r dev.in
pytest-cov==7.0.0
    # via -r dev.in
python-dotenv==1.2.2
    # via
    #   -r base.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r base.in
python-multipart==0.0.22
    # via -r base.in
pytokens==0.4.1
    # via black
pyyaml==6.0.3
    # via uvicorn
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.48
    # via -r base.in
starlette==0.52.1
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
uvicorn[standard]==0.41.0
    # via -r base.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
